[DFT] Add more warning flags
Fix warnings when compiling the DFT domain with some warning flags:
-Wall -Wextra -Wshadow -Wconversion -Wpedantic
The same warnings can be added to more domains if desired.

+++ tests/unit_tests/dft/source/compute_tests.cpp
@@ -39,9 +39,9 @@ extern std::vector<sycl::device *> devices;
 
 namespace {
 
-class ComputeTests : public ::testing::TestWithParam<std::tuple<sycl::device *, std::int64_t>> {};
+class ComputeTests : public ::testing::TestWithParam<std::tuple<sycl::device *, std::size_t>> {};
 
-std::vector<std::int64_t> lengths{ 8, 21, 128 };
+std::vector<std::size_t> lengths{ 8, 21, 128 };
 
 /* test_in_place_buffer() */
 TEST_P(ComputeTests, RealSinglePrecisionInPlaceBuffer) {
